requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
typer>=0.9.0
rich>=13.5.0
pyperclip>=1.8.2
